// types.ts
export enum SriStatus {
  PENDING = 'PENDIENTE',
  RECEIVED = 'RECIBIDA',
  AUTHORIZED = 'AUTORIZADA',
  REJECTED = 'RECHAZADA'
}

export enum PaymentStatus {
  PAID = 'PAGADO',
  PENDING = 'PENDIENTE',
  OVERDUE = 'VENCIDO'
}

export enum UserRole {
  ADMIN = 'ADMINISTRADOR',
  SELLER = 'VENDEDOR',
  ACCOUNTANT = 'CONTADOR'
}

export enum DocumentType {
  INVOICE = '01',
  CREDIT_NOTE = '04',
  DEBIT_NOTE = '05',
  RETENTION = '07',
  REMITTANCE = '06'
}
// ...
export interface InvoiceItem {
  productId: string;
  description: string;
  quantity: number;
  unitPrice: number;
  discount: number;
  taxRate: number;
  total: number;
  imageUrl?: string;
  type: 'FISICO' | 'SERVICIO';
}

export interface AppNotification {
  id: string;
  text: string;
  type: 'success' | 'info' | 'warning' | 'error';
  time: Date;
  read: boolean;
}

export interface PaymentMethod {
  code: string;
  label: string;
}

export interface Document {
  id: string;
  type: DocumentType;
  number: string;
  accessKey: string;
  issueDate: string;
  dueDate?: string;
  entityName: string;
  entityRuc?: string;
  entityEmail?: string;
  entityPhone?: string;
  entityAddress?: string;
  total: number;
  status: SriStatus;
  paymentStatus: PaymentStatus;
  paymentMethod?: string;
  additionalInfo?: string;
  items?: InvoiceItem[];
  source?: 'LOCAL' | 'TIENDA_ONLINE';
  authorizedXml?: string; // XML autorizado por el SRI
  // Para Notas de Crédito
  relatedDocumentNumber?: string; // Factura modificada
  relatedDocumentDate?: string;
  relatedDocumentAccessKey?: string;
  creditNoteReason?: string;

  // Para Retenciones
  taxPeriod?: string;
  sustainingDocType?: string;
  sustainingDocNumber?: string;
  sustainingDocDate?: string;
  sustainingDocTotal?: number;
  retentionTaxes?: RetentionTax[];
}
// ...
export interface RetentionTax {
  code: string; // Código del impuesto (1=Renta, 2=IVA, 6=ISD)
  baseImponible: number;
  percentageCode: string; // Código del porcentaje de retención
  percentage: number; // Porcentaje real (ej: 1, 2, 10, 30, etc)
  taxValue: number; // Valor retenido
  fiscalDocCode: string; // Código del documento fiscal (01=Factura, 04=Nota Crédito, etc)
  fiscalDocNumber: string; // Número del documento
}